Add historical data tabs to /leader and admin/participants
We want admins and leaders to be able to view teams from past years.  The UI will now add tabs for up to three prior years to the admin/participants view and the leader dashboard.

This commit also fixes some email validation issues and adds tests of email validation.  Not all the tests are working as of this commit.