Need to set participants active
If participants don't have a value in the status field, they won't be included in update emails.

The set_participants_active.py script will set the status for all participants that don't have a status value.  It has been run on existing servers, but will be needed to update the Ladner server when it gets updated.